requirements.txt
streamlit==1.29.0
python-pptx==0.6.23
python-docx==1.1.0
